+++ Classes/utility.hpp
@@ -31,12 +31,12 @@ class utility{
     //static sf::Vector2f MaxVector(sf::Vector2f* p1, sf::Vector2f* p2);
     //static sf::Vector2f MaxVector(sf::Vector2f* p1, sf::Vector2f* p2);
     static float toRadiant(float degrees);
-    static float toDegrees(float radiant);
+    static double toDegrees(double radiant);
     static int Proportion(int a, int b, int c = 100);
     static int RandInt(int x, int y = 0);
     static float RandFloat(int x, float y = 0);
     static sf::Vector2f RandVector(int x, int y, int offsetX = 0, int offsetY = 0);
-    
+
     static std::string filenameFromPath(std::string path);
 
     static sf::Vector2f left(drawable* obj);
